pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videoinsight"
version = "0.1.0"
description = "CLI tool for generating comprehensive notes from YouTube videos"
authors = [
    { name = "Vighnesh", email = "your.email@example.com" }
]
requires-python = ">=3.10"
keywords = ["youtube", "transcription", "notes", "markdown", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Multimedia :: Video",
    "Programming Language :: Python :: 3.10"
]
dependencies = [
    "typer>=0.9.0",
    "yt-dlp>=2023.3.4",
    "faster-whisper>=0.6.0",
    "spacy>=3.5.3",
    "pydantic>=2.0.0",
    "rich>=13.3.5",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "sentence-transformers>=2.2.2",
    "keybert>=0.7.0",
    "numpy>=1.24.3",
    "pandas>=2.0.1",
    "prefect>=2.10.18",
    "SQLAlchemy>=2.0.15"
]

[project.urls]
"Source Code" = "https://github.com/Voodels/Summerizer-"

[project.scripts]
videoinsight = "videoinsight.cli.commands:app"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
